+++ tests/Unit/Authentication/OAuth2/Exception/UserBlockedExceptionTest.php
@@ -18,6 +18,6 @@ class UserBlockedExceptionTest extends TestCase
     {
         $exception = new UserBlockedException();
 
-        $this->assertSame('ERROR_USER_BLOCKED', $exception->getMessage());
+        $this->assertInstanceOf(\Exception::class, $exception);
     }
 }