+++ translations/en/oesecuritymodule_lang.php
@@ -38,5 +38,7 @@ $aLang = [
     'ERROR_PASSWORD_STRENGTH_3' => "Strong",
     'ERROR_PASSWORD_STRENGTH_4' => "Very strong",
 
-    'GENERATE_STRONG_PASSWORD' => 'Generate Strong Password'
+    'GENERATE_STRONG_PASSWORD' => 'Generate Strong Password',
+
+    'SIGN_IN_PROVIDER' => 'Sign in with %s',
 ];